+++ integrations/room/src/commonIntegrationTest/kotlin/com/powersync/integrations/room/TestDatabase.kt
@@ -10,9 +10,8 @@ import androidx.room.PrimaryKey
 import androidx.room.Query
 import androidx.room.RoomDatabase
 import androidx.room.RoomDatabaseConstructor
-import com.powersync.db.schema.PendingStatement
-import com.powersync.db.schema.PendingStatementParameter
 import com.powersync.db.schema.RawTable
+import com.powersync.db.schema.RawTableSchema
 import com.powersync.db.schema.Schema
 import kotlinx.coroutines.flow.Flow
 
@@ -47,19 +46,7 @@ abstract class TestDatabase : RoomDatabase() {
             Schema(
                 RawTable(
                     name = "user",
-                    put =
-                        PendingStatement(
-                            "INSERT INTO user (id, name) VALUES (?, ?)",
-                            listOf(
-                                PendingStatementParameter.Id,
-                                PendingStatementParameter.Column("name"),
-                            ),
-                        ),
-                    delete =
-                        PendingStatement(
-                            "DELETE FROM user WHERE id = ?",
-                            listOf(PendingStatementParameter.Id),
-                        ),
+                    schema = RawTableSchema("user"),
                 ),
             )
     }